linux: add morphology
This is based off kernel release 3.4

The versatile_defconfig does not set important options like
PCI and SCSI devices, which the qemu arm-versatile device
needs to read the rootfs disk.

The kernel image is installed to /boot/zImage, since `make install`
on ARM may do anything, ARM systems often have the kernel installed
in flash instead of on the disk, but keeping a copy in /boot is
probably ok.