linux.morph: build versatile-express defconfig
versatile is an ARM9, the AXP builds ARMv7, so the default build
won't generate code that is compatible.

However there exists the versatile express machine in qemu, which
is has a cortex A9, which is a compatible architecture.

The following command line worked
qemu-system-arm -M vexpress-a9 -cpu cortex-a9 -serial stdio -m 512 \
-drive file=rootfs.img,if=sd -kernel linux/arch/arm/boot/zImage \
-append 'root=/dev/mmcblk0p1 console=ttyAMA0 mem=512M rw'
Replace -serial stdio with -nographic if the sdl window isn't wanted.