ARM: Add ARCH_SUPPORTS_BIG_ENDIAN
Add ARCH_SUPPORTS_BIG_ENDIAN define to arch/arm/Kconfig. Since it used
to live in arch/arm/mach-ixp4xx/Kconfig we move it up to the ARM one
as the next best place for it.

It may be that it could be moved to a more generic location, with the
other ARCH_SUPPORT functions.

Signed-off-by: Ben Dooks <ben.dooks@codethink.co.uk>