atags_to_fdt: /memory node needs to take into account #cells and #size
The atags_to_fdt code will either update or create the /memory node and
therefore needs to take into account the root node's settings for the
size and number of cells in the node.

Read the root nodes #cells and #size before creating the list of memory
cells so that they are the size the memory parsing in the early fdt
initialisation code recognises these as valid memory areas.

Currently the default dtb file for the vexpress a15_a7 tile makes the
root node be 2 cells, 2 words per cell which means this code fails to
produce valid memory areas when booting.

Note, we have yet to verify if we can give the /memory node can be given
it's own #size and #cells, but that would invovle fixing the fdt code
which just gets the #cells and #size for the root node.